+++ Tests/Integration/MSFT_AdcsCertificationAuthority.Integration.Tests.ps1
@@ -1,19 +1,10 @@
-# Suppress this PSSA message because we need to allow credentials to be
-# set when running the tests.
+<#
+    Suppress this PSSA message because we need to allow credentials to be
+    set when running the tests.
+#>
 [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingConvertToSecureStringWithPlainText', '')]
 param ()
 
-<#
-    IMPORTANT: To run these tests requires a local Administrator account to be
-    available on the machine running the tests that can be used to install the
-    ADCS component being tested. Please change the following values to the
-    credentials that are set up for this purpose.
-    When these tests are run on AppVeyor, ADCS-Cert-Authority will be installed
-    and a new Administrator account will be created that uses credentials that
-    match the ones following.
-#>
-$script:adminUsername = 'AdcsAdmin'
-$script:adminPassword = 'NotPass12!'
 $script:DSCModuleName = 'ActiveDirectoryCSDsc'
 $script:DSCResourceName = 'MSFT_AdcsCertificationAuthority'
 
@@ -38,6 +29,19 @@ $TestEnvironment = Initialize-TestEnvironment `
 # Using try/finally to always cleanup even if something awful happens.
 try
 {
+    <#
+        IMPORTANT: To run these tests requires a local Administrator account to be
+        available on the machine running the tests that can be used to install the
+        ADCS component being tested. This account will be created automatically for
+        the tests and removed afterward.
+
+        When these tests are run on AppVeyor, ADCS-Cert-Authority will be installed
+        and a new Administrator account will be created that uses credentials that
+        match the ones following.
+    #>
+    $script:adminUsername = 'AdcsAdminTest'
+    $script:adminPassword = ConvertTo-SecureString -String 'NotPass12!' -AsPlainText -Force
+
     # Ensure that the tests can be performed on this computer
     if (-not (Test-WindowsFeature -Name 'ADCS-Cert-Authority'))
     {
@@ -45,24 +49,24 @@ try
         return
     }
 
-    #region Integration Tests
-    $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName)_Install.config.ps1"
-    . $ConfigFile -Verbose -ErrorAction Stop
+    # Create a new Local User in the administrators group
+    $script:adminCredential = New-Object `
+        -TypeName System.Management.Automation.PSCredential `
+        -ArgumentList ($script:adminUsername, $script:adminPassword)
+    New-LocalUserInAdministratorsGroup -User $script:adminUsername -Password $script:adminPassword
 
     Describe "$($script:DSCResourceName)_Install_Integration" {
+        $configFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName)_Install.config.ps1"
+        . $configFile -Verbose -ErrorAction Stop
+
         Context 'Install ADCS Certification Authority' {
-            #region DEFAULT TESTS
             It 'Should compile and apply the MOF without throwing' {
                 {
-                    $secureAdminPassword = ConvertTo-SecureString -String $script:adminPassword -AsPlainText -Force
-                    $adminCred = New-Object -TypeName System.Management.Automation.PSCredential `
-                        -ArgumentList ($script:adminUsername, $secureAdminPassword)
-
-                    $ConfigData = @{
+                    $configData = @{
                         AllNodes = @(
                             @{
                                 NodeName                    = 'localhost'
-                                AdminCred                   = $AdminCred
+                                AdminCred                   = $script:adminCredential
                                 PsDscAllowPlainTextPassword = $true
                             }
                         )
@@ -70,7 +74,7 @@ try
 
                     & "$($script:DSCResourceName)_Install_Config" `
                         -OutputPath $TestDrive `
-                        -ConfigurationData $ConfigData
+                        -ConfigurationData $configData
 
                     Start-DscConfiguration `
                         -Path $TestDrive `
@@ -85,7 +89,6 @@ try
             It 'Should be able to call Get-DscConfiguration without throwing' {
                 { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should -Not -Throw
             }
-            #endregion
 
             It 'Should have set the resource and all the parameters should match' {
                 $current = Get-DscConfiguration | Where-Object {
@@ -96,23 +99,18 @@ try
         }
     }
 
-    $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName)_Uninstall.config.ps1"
-    . $ConfigFile -Verbose -ErrorAction Stop
-
     Describe "$($script:DSCResourceName)_Uninstall_Integration" {
+        $configFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName)_Uninstall.config.ps1"
+        . $configFile -Verbose -ErrorAction Stop
+
         Context 'Uninstall ADCS Certification Authority' {
-            #region DEFAULT TESTS
             It 'Should compile and apply the MOF without throwing' {
                 {
-                    $secureAdminPassword = ConvertTo-SecureString -String $script:adminPassword -AsPlainText -Force
-                    $adminCred = New-Object -TypeName System.Management.Automation.PSCredential `
-                        -ArgumentList ($script:adminUsername, $secureAdminPassword)
-
-                    $ConfigData = @{
+                    $configData = @{
                         AllNodes = @(
                             @{
                                 NodeName                    = 'localhost'
-                                AdminCred                   = $AdminCred
+                                AdminCred                   = $script:adminCredential
                                 PsDscAllowPlainTextPassword = $true
                             }
                         )
@@ -120,7 +118,7 @@ try
 
                     & "$($script:DSCResourceName)_Uninstall_Config" `
                         -OutputPath $TestDrive `
-                        -ConfigurationData $ConfigData `
+                        -ConfigurationData $configData `
                         -ErrorAction Stop
 
                     Start-DscConfiguration `
@@ -136,7 +134,6 @@ try
             It 'Should be able to call Get-DscConfiguration without throwing' {
                 { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should -Not -Throw
             }
-            #endregion
 
             It 'Should have set the resource and all the parameters should match' {
                 $current = Get-DscConfiguration | Where-Object {
@@ -146,11 +143,11 @@ try
             }
         }
     }
-    #endregion
 }
 finally
 {
     #region FOOTER
+    Remove-LocalUser -Name $script:adminUsername -ErrorAction SilentlyContinue
     Restore-TestEnvironment -TestEnvironment $TestEnvironment
     #endregion
 }